Update merge_halves (among other things)
Merge_halves updates:
Because copy_to creates a temporary buffer on each invocation, and
because said buffer is of the same block of memory each time, creating a
buffer to hold the first block significantly reduces the number of
allocs, frees, and copies in any cases where (in shorthand) two early
smallest values are from the second block.

is_sorted:
This check doesn't really add that much to the length of the function
according to some basic benchmarking I did, and can improve speed
significantly.

test updates:
With the modified merge_halves, I needed to formalize a test to guard
against accidental double drops etc, so a test with std::String was
added.